+++ src/renderer/Components/App.tsx
@@ -1,25 +1,19 @@
 // import basic style
 import './App.scss';
 // import packages
-import {RefObject, useEffect, useMemo, useRef, useState} from 'react';
+import React, {RefObject, useEffect, useMemo, useRef, useState} from 'react';
 import {WebviewTag} from 'electron';
-import {motion} from 'framer-motion';
 // import modules
 import {ipcUserData, ipcWindowManager} from './RendererIpcHandler';
 // import components
 import TitleBar from './TitleBar/TitleBar';
 import SideBar from './SideBar/SideBar';
-import Sda1WebUi from './WebuiCard/Sda1WebUi';
 import StatusContext, {StatusContextType} from './GlobalStateContext';
 import BlockBackground from '../Windows/BlockBackground';
 import WebUiViewer from './Launcher/WebUiViewer';
-import {RendererLogError, WebuiList} from '../../AppState/AppConstants';
-import SdLshWebUi from './WebuiCard/SdLshWebUi';
+import {RendererLogError, sideBarButtonId, WebuiList} from '../../AppState/AppConstants';
 import {AppConfig} from '../../AppState/InterfaceAndTypes';
-import OOBABOOGAWebUI from './WebuiCard/OOBABOOGAWebUI';
-
-const sideBarPictureId: string = 'sideBarPicture';
-const sideBarTextId: string = 'sideBarText';
+import PageContainer from './PageContainer';
 
 export default function App() {
   // Whether user theme is on dark mode
@@ -27,8 +21,8 @@ export default function App() {
 
   // Block background of an opened window to be not interactive
   const [blockBackground, setBlockBackground] = useState<boolean>(false);
-  const [selectedPage, setSelectedPage] = useState<string>(sideBarPictureId);
-  const [installedWebUi, setInstalledWebUi] = useState<WebuiList>({AUTOMATIC1111: false, LSHQQYTIGER: false, OOBABOOGA: false});
+  const [selectedPage, setSelectedPage] = useState<number>(sideBarButtonId.Image);
+  const [installedWebUi, setInstalledWebUi] = useState<WebuiList>({AUTOMATIC1111: false, LSHQQYTIGER: false, OOBABOOGA: false, RSXDALV: false});
   const [webuiRunning, setWebuiRunning] = useState<boolean>(false);
 
   const webViewRef = useRef<WebviewTag>(null);
@@ -64,6 +58,7 @@ export default function App() {
           AUTOMATIC1111: value.WebUi.AUTOMATIC1111?.installed,
           LSHQQYTIGER: value.WebUi.LSHQQYTIGER?.installed,
           OOBABOOGA: value.WebUi.OOBABOOGA?.installed,
+          RSXDALV: value.WebUi.RSXDALV?.installed,
         });
         return value;
       })
@@ -91,7 +86,7 @@ export default function App() {
 
   /* (Main Component)
    *
-   * [MainWindowBlock] Blocking background interactive when showing a window (Like installing webui)
+   * [BlockBackground] Blocking background interactive when showing a window (Like installing webui)
    * Using props to showing background block because of AnimatePresence for enter and exit animation
    *
    * [TitleBar] Main app title bar
@@ -101,7 +96,7 @@ export default function App() {
    * [webuiRunning] If webui is running, show {WebUiViewer} to show webui browser like
    * If webui is not running showing install or run webui
    *
-   * [InstallSD] The window component for installing specif webui
+   * [PageContainer] Contain sidebar pages (Image Generation and ...)
    *
    */
 
@@ -111,44 +106,7 @@ export default function App() {
       <TitleBar />
       <SideBar />
       {/* Webui cards container */}
-      <div className="absolute bottom-0 left-24 right-0 top-10 overflow-hidden">
-        {webuiRunning ? (
-          <WebUiViewer />
-        ) : (
-          <>
-            <motion.div
-              className="absolute flex h-full w-full flex-wrap content-start opacity-0"
-              animate={{
-                translateY: selectedPage === sideBarPictureId ? '0px' : '-700px',
-                opacity: selectedPage === sideBarPictureId ? 1 : 0,
-              }}
-              transition={{
-                ease: selectedPage === sideBarPictureId ? 'backOut' : 'backIn',
-                duration: 0.3,
-                delay: selectedPage === sideBarPictureId ? 0.15 : 0,
-              }}>
-              {/* Automatic1111 WebUi */}
-              <Sda1WebUi />
-              {/* LSHQQYTIGER (DirectMl) WebUi */}
-              <SdLshWebUi />
-            </motion.div>
-            <motion.div
-              className="absolute flex h-full w-full flex-wrap content-start opacity-0"
-              animate={{
-                translateY: selectedPage === sideBarTextId ? '0px' : '700px',
-                opacity: selectedPage === sideBarTextId ? 1 : 0,
-              }}
-              transition={{
-                ease: selectedPage === sideBarTextId ? 'backOut' : 'backIn',
-                duration: 0.3,
-                delay: selectedPage === sideBarTextId ? 0.15 : 0,
-              }}>
-              {/* OOBABOOGAWebUI WebUi */}
-              <OOBABOOGAWebUI />
-            </motion.div>
-          </>
-        )}
-      </div>
+      <div className="absolute bottom-0 left-24 right-0 top-10 overflow-hidden">{webuiRunning ? <WebUiViewer /> : <PageContainer />}</div>
     </StatusContext.Provider>
   );
 }