+++ src/renderer/Components/SideBar/SideBarButton.tsx
@@ -4,7 +4,7 @@ import {useContext} from 'react';
 import {motion, Variants} from 'framer-motion';
 // Import Components
 import StatusContext, {StatusContextType} from '../GlobalStateContext';
-import {getBlack, getLynxBlue, getWhite, getWhiteFourth} from '../../../AppState/AppConstants';
+import {getSidebarButtonId, getWhite, getWhiteFourth} from '../../../AppState/AppConstants';
 
 type Props = {
   // Extra class names for the root element
@@ -14,9 +14,9 @@ type Props = {
   // Button id
   btnId: string;
   // Currently selected button (id)
-  selected: string;
+  selected: number;
   // Set the currently selected button (id)
-  setSelected: (id: string) => void;
+  setSelected: (id: number) => void;
 };
 
 export default function SideBarButton({extraClasses, icon, btnId, selected, setSelected}: Props) {
@@ -24,14 +24,12 @@ export default function SideBarButton({extraClasses, icon, btnId, selected, setS
 
   const HandleClick = () => {
     // Set currently page selected
-    setSelected(btnId);
+    setSelected(getSidebarButtonId(btnId));
   };
 
   // Motion animation variants
   const buttonVariants: Variants = {
     hover: {
-      borderWidth: '1px',
-      borderColor: isDarkMode ? getWhite(0.2) : getBlack(0.2),
       backgroundColor: isDarkMode ? getWhite(0.1) : getWhiteFourth(),
       transition: {duration: 0.15},
     },
@@ -42,14 +40,10 @@ export default function SideBarButton({extraClasses, icon, btnId, selected, setS
       transition: {duration: 0.1},
     },
     active: {
-      borderWidth: '2px',
-      borderColor: getLynxBlue(0.5),
       backgroundColor: isDarkMode ? getWhite(0) : getWhiteFourth(0),
       transition: {duration: 0.3},
     },
     deActive: {
-      borderWidth: '0px',
-      borderColor: 'rgba(0,0,0,0)',
       transition: {duration: 0.3},
     },
   };
@@ -61,13 +55,16 @@ export default function SideBarButton({extraClasses, icon, btnId, selected, setS
       variants={buttonVariants}
       whileTap="tap"
       whileHover="hover"
-      animate={selected === btnId ? 'active' : 'deActive'}
+      animate={selected === getSidebarButtonId(btnId) ? 'active' : 'deActive'}
       onClick={HandleClick}
       className={['relative mt-[0.5rem] flex h-[4.4rem] w-[4.4rem] cursor-default items-center justify-center rounded-3xl', extraClasses].join(' ')}>
       {/* Button Icon */}
       <img
         src={icon}
-        className={[selected === btnId ? 'LynxBlueFilter' : 'imgDarkLightFilter', 'pointer-events-none h-8 w-8 opacity-[85%]'].join(' ')}
+        className={[
+          selected === getSidebarButtonId(btnId) ? 'LynxBlueFilter' : 'imgDarkLightFilter',
+          'pointer-events-none h-8 w-8 opacity-[85%] transition-all duration-300',
+        ].join(' ')}
         alt={btnId}
       />
     </motion.button>