+++ src/renderer/Components/WebuiCard/WebUI/SdLshWebUi.tsx
@@ -2,16 +2,16 @@
 import React, {useEffect, useState} from 'react';
 import {AnimatePresence} from 'framer-motion';
 // Import components
-import CardWebUi from './Card/CardWebUi';
+import CardWebUi from '../Card/CardWebUi';
+import WebUiSDLaunchSettings from '../LaunchSettings/WebUiSDLaunchSettings';
 // Import assets
-import oobaboogaImg from '../../../Assets/AiCard/RepoUser/OOBABOOGA.png';
-import cardBackgroundImg from '../../../Assets/AiCard/BGCardTextGeneration.jpg';
-import {ipcUserData} from '../RendererIpcHandler';
-import WebUiTGLaunchSettings from './WebUiTGLaunchSettings';
+import lshqqytigerImg from '../../../../Assets/AiCard/RepoUser/LSHQQYTIGER.png';
+import stableDiffusionImg from '../../../../Assets/AiCard/BGCardImageGeneration.png';
+import {ipcUserData} from '../../RendererIpcHandler';
 
-const userName: string = 'OOBABOOGA';
+const userName: string = 'LSHQQYTIGER';
 
-export default function OOBABOOGAWebUI() {
+export default function SdLshWebUi() {
   const [launchSettingsMenu, setLaunchSettingsMenu] = useState(false);
 
   // Toggle to show launch settings or WebUi card
@@ -29,16 +29,16 @@ export default function OOBABOOGAWebUI() {
       <CardWebUi
         webuiData={{
           repoUserName: userName,
-          webUiDesc: 'Text Generation WebUi',
-          repoAvatarImg: oobaboogaImg,
-          cardBackgroundImg,
-          cardBackgroundPosition: 'object-bottom',
+          webUiDesc: 'Stable Diffusion DirectML WebUi',
+          repoAvatarImg: lshqqytigerImg,
+          cardBackgroundImg: stableDiffusionImg,
+          cardBackgroundPosition: 'object-right-top',
         }}
         launchSettingsMenu={launchSettingsMenu}
         toggleSettings={ToggleSettings}
       />
       {/* AnimatePresence for exit animations */}
-      <AnimatePresence>{launchSettingsMenu && <WebUiTGLaunchSettings repoUserName={userName} ToggleSettings={ToggleSettings} />}</AnimatePresence>
+      <AnimatePresence>{launchSettingsMenu && <WebUiSDLaunchSettings repoUserName={userName} ToggleSettings={ToggleSettings} />}</AnimatePresence>
     </>
   );
 }